+++ internal/transport/transport.go
@@ -1,4 +1,4 @@
-package transfer
+package transport
 
 import (
 	"context"
